feat: attempt to wrap in navigator
