+++ src/screens/Bookmarks/index.tsx
@@ -1,7 +1,6 @@
-import { RouteProp, useNavigation, useRoute } from '@react-navigation/native'
-import { StackNavigationProp } from '@react-navigation/stack'
 import React from 'react'
-import { Alert } from 'react-native'
+import { RouteProp, useRoute } from '@react-navigation/native'
+import { createStackNavigator } from '@react-navigation/stack'
 
 import { BookmarksList } from '../../components/Bookmarks'
 import Container from '../../components/Container'
@@ -11,31 +10,19 @@ import { NavigationParams } from '../../types/navigation'
 import Navbar from './Navbar'
 
 type Route = RouteProp<NavigationParams, Screens.Bookmarks>
-type Navigation = StackNavigationProp<NavigationParams, Screens.Bookmarks>
+
+const { Screen, Navigator } = createStackNavigator()
 
 const BookmarksScreen = () => {
   const route = useRoute<Route>()
-  const navigation = useNavigation<Navigation>()
 
   const { folderData } = route.params
 
-  const onItemPress = ( isFolder: boolean, name: string ) => {
-    if ( isFolder ) {
-      navigation.push( Screens.Bookmarks,
-        {
-          folderData:
-          folderData.find( ( folder ) => folder.name === name )?.bookmarks
-          || folderData,
-          currentFolder: name,
-        } )
-    } else {
-      Alert.alert( `you clicked on ${name}` )
-    }
-  }
-
   return (
     <Container>
-      <BookmarksList data={folderData} onItemPress={onItemPress} />
+      <Navigator>
+        <Screen name={Screens.Bookmarks} component={BookmarksList} initialParams={folderData} />
+      </Navigator>
     </Container>
   )
 }
